+++ index.html
@@ -234,5 +234,8 @@
 
 </body>
 
-<script src="js/main.js"></script>
+
+<script src="js/util.js" type="module"></script>
+<script src="js/data.js" type="module"></script>
+<script src="js/main.js" type="module"></script>
 </html>
